# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fresher Resume Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes/{resumeId}/personal-info:
    get:
      operationId: getPersonalInfo
      tags: [sections]
      summary: Get personal info for a resume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personal info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalInfo"
    put:
      operationId: upsertPersonalInfo
      tags: [sections]
      summary: Create or update personal info
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalInfoInput"
      responses:
        "200":
          description: Updated personal info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalInfo"

  # --- Career Objective ---
  /resumes/{resumeId}/objective:
    get:
      operationId: getObjective
      tags: [sections]
      summary: Get career objective
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"
    put:
      operationId: upsertObjective
      tags: [sections]
      summary: Create or update career objective
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectiveInput"
      responses:
        "200":
          description: Updated objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"

  # --- Education ---
  /resumes/{resumeId}/education:
    get:
      operationId: listEducation
      tags: [sections]
      summary: List education entries
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Education list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Education"
    post:
      operationId: createEducation
      tags: [sections]
      summary: Add education entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "201":
          description: Created education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
  /resumes/{resumeId}/education/{educationId}:
    patch:
      operationId: updateEducation
      tags: [sections]
      summary: Update education entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: educationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "200":
          description: Updated education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
    delete:
      operationId: deleteEducation
      tags: [sections]
      summary: Delete education entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: educationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- Skills ---
  /resumes/{resumeId}/skills:
    get:
      operationId: listSkills
      tags: [sections]
      summary: List skills
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [sections]
      summary: Add skill
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /resumes/{resumeId}/skills/{skillId}:
    patch:
      operationId: updateSkill
      tags: [sections]
      summary: Update skill
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [sections]
      summary: Delete skill
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: skillId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- Projects ---
  /resumes/{resumeId}/projects:
    get:
      operationId: listProjects
      tags: [sections]
      summary: List projects
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [sections]
      summary: Add project
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /resumes/{resumeId}/projects/{projectId}:
    patch:
      operationId: updateProject
      tags: [sections]
      summary: Update project
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [sections]
      summary: Delete project
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- Experience ---
  /resumes/{resumeId}/experience:
    get:
      operationId: listExperience
      tags: [sections]
      summary: List experience entries
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experience list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
    post:
      operationId: createExperience
      tags: [sections]
      summary: Add experience entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /resumes/{resumeId}/experience/{experienceId}:
    patch:
      operationId: updateExperience
      tags: [sections]
      summary: Update experience entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: experienceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
    delete:
      operationId: deleteExperience
      tags: [sections]
      summary: Delete experience entry
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: experienceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- Certifications ---
  /resumes/{resumeId}/certifications:
    get:
      operationId: listCertifications
      tags: [sections]
      summary: List certifications
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certification"
    post:
      operationId: createCertification
      tags: [sections]
      summary: Add certification
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationInput"
      responses:
        "201":
          description: Created certification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
  /resumes/{resumeId}/certifications/{certificationId}:
    patch:
      operationId: updateCertification
      tags: [sections]
      summary: Update certification
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: certificationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationInput"
      responses:
        "200":
          description: Updated certification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
    delete:
      operationId: deleteCertification
      tags: [sections]
      summary: Delete certification
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
        - name: certificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # --- Languages ---
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    PersonalInfo:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedin:
          type: ["string", "null"]
        portfolio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
      required:
        - id
        - resumeId
        - fullName
        - email
        - phone
    PersonalInfoInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedin:
          type: string
        portfolio:
          type: string
        address:
          type: string
        photoUrl:
          type: string
      required:
        - fullName
        - email
        - phone
    Objective:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        summaryText:
          type: string
      required:
        - id
        - resumeId
        - summaryText
    ObjectiveInput:
      type: object
      properties:
        summaryText:
          type: string
          maxLength: 500
      required:
        - summaryText
    Education:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        institution:
          type: string
        degree:
          type: string
        fieldOfStudy:
          type: string
        graduationYear:
          type: integer
        cgpa:
          type: string
      required:
        - id
        - resumeId
        - institution
        - degree
        - fieldOfStudy
        - graduationYear
        - cgpa
...